requirements.txt
asttokens==3.0.0
babel==2.17.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.2
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
debugpy==1.8.14
decorator==5.2.1
executing==2.2.0
fonttools==4.58.0
future==1.0.0
geocoder==1.38.1
idna==3.10
iniconfig==2.1.0
ipykernel==6.29.5
ipython==9.2.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
matplotlib==3.10.3
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.2.5
packaging==25.0
pandas==2.2.3
parso==0.8.4
patsy==1.0.1
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
prompt_toolkit==3.0.51
psutil==7.0.0
pure_eval==0.2.3
Pygments==2.19.1
pyparsing==3.2.3
pyperclip==1.9.0
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pywin32==310
pyzmq==26.4.0
ratelim==0.1.6
requests==2.32.3
scikit-learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
stack-data==0.6.3
statsmodels==0.14.4
threadpoolctl==3.6.0
tk==0.1.0
tkcalendar==1.6.1
tkintermapview==1.29
tornado==6.5
traitlets==5.14.3
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
